fix: GUI responsiveness when changing tracks
Refactor KNotification to use QImage instead of tossing around QPixmaps.
This reduces unnessecary serialization. In addition, resize notification
thumbnail to 512px.

The GUI might still take too long to update for extremely large images,
as resizing is not free. This is still an improvement.

In the future, we could cache these smaller thumbnails, or on Linux,
pass the `image-path` parameter.