fix: fill tags from filename if tag is empty in metadata
`Song::guessTags()` currently does not run if any metadata is available,
even if it can extract info unavailable in the metadata. This leads to
song title being set to Unknown if metadata is available but no title is
available in the metadata. This commit makes guessTags() always run, and
only fill in what is unset at the end. The result is that it merges the
metadata and the information available from the filename with priority
given to the metadata.