3rdparty/solid-lite: restore
This is a modified copy of KDE Solid inherited from upstream Cantata.
While Solid is definitely available as a package on Linux, it might be
harder to install KDE Frameworks on Windows and Linux. Keep this in tree
for now until KF6::Solid can be correctly vendored and implemented.

Also, the devices feature might not even be working properly right now.